chore: prepare 0.0.14 release (#28)
* ci: switch TestPyPI publish job to OIDC trusted publisher

* chore(deps): bump the uv group across 1 directory with 2 updates (#26)

Bumps the uv group with 2 updates in the / directory: [torch](https://github.com/pytorch/pytorch) and [msgpack](https://github.com/msgpack/msgpack-python).


Updates `torch` from 2.11.0 to 2.12.1
- [Release notes](https://github.com/pytorch/pytorch/releases)
- [Changelog](https://github.com/pytorch/pytorch/blob/main/RELEASE.md)
- [Commits](https://github.com/pytorch/pytorch/compare/v2.11.0...v2.12.1)

Updates `msgpack` from 1.1.2 to 1.2.1
- [Release notes](https://github.com/msgpack/msgpack-python/releases)
- [Changelog](https://github.com/msgpack/msgpack-python/blob/main/CHANGELOG.md)
- [Commits](https://github.com/msgpack/msgpack-python/compare/v1.1.2...v1.2.1)

---
updated-dependencies:
- dependency-name: torch
  dependency-version: 2.12.1
  dependency-type: direct:production
  dependency-group: uv
- dependency-name: msgpack
  dependency-version: 1.2.1
  dependency-type: indirect
  dependency-group: uv
...


* chore(deps): bump actions/checkout from 6 to 7 (#27)

Bumps [actions/checkout](https://github.com/actions/checkout) from 6 to 7.
- [Release notes](https://github.com/actions/checkout/releases)
- [Changelog](https://github.com/actions/checkout/blob/main/CHANGELOG.md)
- [Commits](https://github.com/actions/checkout/compare/v6...v7)

---
updated-dependencies:
- dependency-name: actions/checkout
  dependency-version: '7'
  dependency-type: direct:production
  update-type: version-update:semver-major
...


* perf: cache loaded models/pipelines

* test: add cache-hit regression tests for model_cache

* chore: prepare 0.0.14 release

* chore(deps): update all dependencies to their latest compatible versions

* fix: reconfigure stdout/stderr to UTF-8

* chore(deps): bump pyproject.toml

* Guard stdout/stderr reconfigure() with getattr for robustness

---------

+++ src/Jabberjay/Models/Spectra0/model.py
@@ -27,7 +27,7 @@ class Bottle2neck(nn.Module):
         self, inplanes, planes, kernel_size: int = 3, dilation: int = 1, scale=8
     ):
         super().__init__()
-        width = int(math.floor(planes / scale))
+        width = math.floor(planes / scale)
         self.conv1 = nn.Conv1d(inplanes, width * scale, kernel_size=1)
         self.bn1 = nn.BatchNorm1d(width * scale)
         self.nums = scale - 1
